Pipelines/requirements.txt
# OpenWebUI Pipeline Dependencies

# Core Pipeline dependencies (HTTP client mode)
pydantic>=2.0.0
fastapi>=0.100.0
requests>=2.31.0
aiohttp>=3.8.0
tenacity>=8.4.1
python-dotenv>=1.0.0

# Server dependencies (for local RAG server)
uvicorn>=0.20.0
numpy>=1.24.0
openai>=1.0.0
litellm>=1.0.0
rank-bm25>=0.2.2
faiss-cpu>=1.7.4
tiktoken>=0.5.0
together>=0.2.0

# Document processing (for local server)
PyPDF2>=3.0.0
pandas>=1.5.0

# Future evaluation dependencies (for Phase 2)
# deepeval>=0.20.0
# ragas>=0.1.0

# Additional utilities
openpyxl>=3.0.0